ml/requirements.txt
# LLMrecon ML/AI Requirements
# Python 3.8+ required

# Core ML Libraries
torch>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Deep Learning & NLP
transformers>=4.30.0
gym>=0.26.0
stable-baselines3>=2.0.0

# Visualization & Dashboard
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing (for multi-modal)
Pillow>=10.0.0
opencv-python>=4.8.0
qrcode>=7.4.0

# Data Processing
scipy>=1.11.0
networkx>=3.1

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional GPU Support
# Uncomment based on your CUDA version:
# torch==2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# tensorflow[and-cuda]

# Development Dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0